requirements.txt
# requirements.txt

fastapi
uvicorn
sqlalchemy
pydantic
python-telegram-bot
requests
python-multipart
passlib[bcrypt]
python-jose
python-dotenv